Improve DatePicker API to switch from month to year view
A method to switch between month selector or year selector doesn't
feel declarative to me.

Add a property to change the view.

At the moment the user has to change the state but I would prefer
an enum property, but this requires writing C++ code while
pickers are pure QML code.

Issue: #165